+++ DemoApplication/VisitorApplication/src/SwitchCyclicAcyclicRepository.h
@@ -8,9 +8,9 @@
 #ifndef SWITCHCYCLICACYCLICREPOSITORY_H_
 #define SWITCHCYCLICACYCLICREPOSITORY_H_
 
-#include <Visitor/DefaultLoggingPolicy.h>
-#include <visitor/Typelist.h>
-#include <visitor/BaseKind.h>
+#include <Visitor/DefaultPolicy.h>
+//#include <visitor/Typelist.h>
+//#include <visitor/BaseKind.h>
 #include <Visitor/Cyclic/Repository.h>
 #include <Visitor/Acyclic/Repository.h>
