+++ VisitorFramework/include/Visitor/DefaultPolicy.h
